+++ src/main.js
@@ -4,6 +4,7 @@ const { context } = require('@actions/github')
 const file = require('./file')
 const gpg = require('./gpg')
 const inputs = require('./inputs')
+const outputs = require('./outputs')
 const Repository = require('./repository')
 const transforms = require('./transforms')
 
@@ -15,6 +16,7 @@ const run = async () => {
         }
         info(`Working directory: ${cwd}`)
 
+        // Inputs
         const { owner, repo: repoName } = context.repo
         const {
             token,
@@ -33,9 +35,11 @@ const run = async () => {
             labels,
         } = inputs.parse()
 
+        // Authenticate
         const repo = new Repository(owner, repoName, token)
         await repo.authenticate(commitAuthorName, commitAuthorEmail)
 
+        // Setup GPG
         if (gpgPrivateKey) {
             if (!gpgPassphrase) {
                 throw new Error('No GPG passphrase specified')
@@ -47,20 +51,27 @@ const run = async () => {
             await repo.setupGpg(keyId, gpgProgramFilePath)
         }
 
+        // Define outputs that hasn't been defined yet
+        const currentYear = new Date().getFullYear()
+        let pullRequestNumber = null
+        let pullRequestUrl = null
+
+        // Print current year
+        info(`Current year is "${currentYear}"`)
+
+        // Checkout branch
         const branchExists = await repo.branchExists(branchName)
         info(`Checkout ${branchExists ? 'existing' : 'new'} branch named "${branchName}"`)
         await repo.checkoutBranch(branchName, !branchExists)
 
+        // Search for files to update
         const files = await file.search(path)
         if (files.length === 0) {
             throw new Error(`Found no files matching the path "${singleLine(path)}"`)
         }
-
         info(`Found ${files.length} file(s) matching the path "${singleLine(path)}"`)
 
-        const currentYear = new Date().getFullYear()
-        info(`Current year is "${currentYear}"`)
-
+        // Update files
         for (const file of files) {
             const relativeFile = file.replace(cwd, '.')
             const content = await repo.readFile(file)
@@ -78,14 +89,21 @@ const run = async () => {
             return
         }
 
+        // Stage and commit changes
         await repo.stageWrittenFiles()
 
         const commitMessage = commitBody ? `${commitTitle}\n\n${commitBody}` : commitTitle
         await repo.commit(commitMessage)
         await repo.push()
 
-        const hasPullRequest = await repo.hasPullRequest(branchName)
-        if (!hasPullRequest) {
+        // Create pull request if it hasn't already been created
+        let pullRequest = await repo.getPullRequest(branchName)
+        if (pullRequest) {
+            info(`Pull request ${pullRequest.number} with title "${pullRequest.title}" has already been created`)
+            pullRequestNumber = pullRequest.number
+            pullRequestUrl = pullRequest.html_url
+        } else {
+            // Create pull request
             info(`Create new pull request with title "${pullRequestTitle}"`)
             const createPullRequestResponse = await repo.createPullRequest(
                 branchName,
@@ -93,16 +111,28 @@ const run = async () => {
                 pullRequestBody
             )
 
+            if (createPullRequestResponse.status !== 201) {
+                throw new Error(`Failed to create pull request: ${createPullRequestResponse.status}`)
+            }
+
+            pullRequestNumber = createPullRequestResponse.data.number
+            pullRequestUrl = createPullRequestResponse.data.html_url
+
+            // Add assignees
             if (assignees.length > 0) {
                 info(`Add assignees to pull request: ${JSON.stringify(assignees)}`)
                 await repo.addAssignees(createPullRequestResponse.data.number, assignees)
             }
 
+            // Add labels
             if (labels.length > 0) {
                 info(`Add labels to pull request: ${JSON.stringify(labels)}`)
                 await repo.addLabels(createPullRequestResponse.data.number, labels)
             }
         }
+
+        // Set outputs
+        outputs.set(currentYear, branchName, pullRequestNumber, pullRequestUrl)
     } catch (err) {
         // @ts-ignore
         setFailed(err.message)